CFE-2965 Add link reference to sys.uqhost
sys.fqhost links to sys.uqhost, but the link is broken. Probably because it
falls later on the same page and it isn't properly resolved. We have had to do
similar for sys.policy_entry_dirname and sys.policy_entry_filename in the past.

This change adds a reference link for sys.uqhost so that it will be sure to
resolve.

(cherry picked from commit 78ab37ef4bee86008a662912f9afb17a9276bb22)